feat: заменил в Color конструктор с аргументом std::uint8_t на std::uint32_t, и теперь инициализирует поле .rgba. Удалил метод рендер из интерфейса Drawable, так как с ним неудобно работать с глубиной

+++ raytracer/graphics/color.hpp
@@ -4,7 +4,7 @@
 namespace gfx {
     struct Color {
         Color() = default;
-        explicit Color(std::uint8_t value);
+        explicit Color(std::uint32_t value);
         Color(const math::vec4f& vec);
         Color(const math::vec4ub& vec);
         Color(std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a = 255);
